Automate image build, refactor Dockerfile
This commits adds a Github action to automage image build
and push to Harbor. The action is triggered on every push to the
main branch and on every PR.
The Dockerfile is refactored to use Alpine base images in order
to reduce the image size and improve security.
Also, the repo was renamed from `esp32-sota` to `esp32-flashjob`,
so any reference to `esp32-sota` in the codebase was updated
to `esp32-flashjob`.

Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>